app/requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.1
Django==4.2
django-cors-headers==3.14.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
